+++ ContextDesktop/src/main/java/com/github/migi_1/Context/damageDealers/StaticObstacle.java
@@ -1,4 +1,4 @@
-package obstacles;
+package com.github.migi_1.Context.damageDealers;
 
 import com.jme3.asset.AssetManager;
 import com.jme3.collision.CollisionResults;
